Model annotation refractoring and minor additions
