+++ Frontend/Pages/UserManagement/Login.razor
@@ -50,7 +50,7 @@
 </div>
 
 @code {
-    readonly Models.Login _login = new();
+    readonly Models.DTO.Login _login = new();
 
     private void HandleValidSubmit() {
         throw new NotImplementedException();
