--- Models/Login.cs
@@ -1,13 +1,0 @@
-﻿using System.ComponentModel.DataAnnotations;
-
-namespace Models; 
-
-public class Login {
-    
-    [Required]
-    [EmailAddress]
-    public string Email { get; set; } = null!;
-
-    [Required]
-    public string Password { get; set; } = null!;
-}
